fold(CSprite): dissolve the view onto the real CDDrawWorker; slot 14 becomes its own virtual
Stage 4 of the 3-way fold: CSprite IS CDDrawWorker (vtbl 0x1efbe8). The view is gone -
it survives only as a typedef, so the ~177 call sites keep reading in their own domain
language (the tree's own CMoviePlayer precedent: CDDScreen/CDDPageMgr are typedefs of
it) while there is ONE type, ONE layout and ONE vtable behind them.

What the fold retires:
  * CFrameArray (the +0x10 "frame CObArray" view) is DELETED - it was the real MFC
    ::CObArray all along, and the view itself proved it: CSprite::InsertFrame already
    cast its own array to `(CObArray*)` to call SetAtGrow @0x1b5822, while
    CDDrawWorker::AddFrameAt calls m_items.SetAtGrow natively. Same object, two
    spellings. Frame access now uses the real API - (CImage*)x->m_items.GetAt(i) /
    GetSize() - because MFC's m_pData/m_nSize are protected; that is the spelling
    WwdGameObject.cpp already matched with.
  * Slot 14 @0x151f00 was declared as ?InsertFrame@CSprite@@ while BEING this vtable's
    slot-14 body; it is now CDDrawWorker's real virtual, which RETIRES that WIRING row
    (the surviving 11/12/13 rows are CImageSet's - stage 5). Its return type is CImage*,
    not the `i32` the slot was declared with: the body's own mangled name says PAVCImage.
    The two registry forwarders hand that pointer back through their i32 signatures.
  * +0x0c: CLoadable models m_0c as a plain i32 because each derivative owns a DIFFERENT
    parent type, so retyping the shared base would be wrong for the others. The concrete
    CImageParent* lives in ONE typed accessor on the leaf (Owner()/SetOwner()) instead of
    a cast at every site - and a prior lane's note in WwdGameObject.cpp predicted exactly
    this ("type that member and this falls out"). It did: the (CImageParent*)m_c cast is gone.

The only casts added are the authentic ones: MFC's CObArray genuinely stores CObject*,
so the element's real type needs the downcast at the sites that consume it as CImage*.
cl was the oracle for WHERE - a mechanical m_pData[i] -> GetAt(i) rewrite, then let the
compiler name the ~14 sites that actually needed the cast, rather than guessing them in.

Bytes: the fold GAINED, it did not cost - exact 2395 -> 2397, fuzzy 73.39 held (floor
2395 / high-water 73.39). Slot-binding: my fold clears exactly ONE row (slot 14); the
file also absorbs 9 rows other lanes had already fixed in code but whose baseline rows
lingered, so the re-derived total is 111 -> 101 (main's 102 minus mine). Re-derived by
the tool, not ported. All FATAL gates green.

(cherry picked from commit aa43d572e0ad32202eaf249daaf29a994b939ed5)

+++ src/DDrawMgr/DDrawWorkerRegistry.cpp
@@ -132,7 +132,7 @@ i32 CDDrawWorkerRegistry::DispatchKeyed38(void* rec, const char* key, i32 a3, i3
     if (worker == 0) {
         return 0;
     }
-    return worker->InsertFrame(rec, a3, a4);
+    return (i32)worker->InsertFrame(rec, a3, a4); // slot 14 returns CImage*; forwarded as i32
 }
 
 // ---------------------------------------------------------------------------
@@ -177,7 +177,7 @@ i32 CDDrawWorkerRegistry::Forward34(i32 a1, CDDrawWorker* worker, i32 a3, i32 a4
 
 RVA(0x00154f20, 0x1b)
 i32 CDDrawWorkerRegistry::Forward38(void* rec, CDDrawWorker* worker, i32 a3, i32 a4) {
-    return worker->InsertFrame(rec, a3, a4);
+    return (i32)worker->InsertFrame(rec, a3, a4); // slot 14 returns CImage*; forwarded as i32
 }
 
 RVA(0x00154f40, 0x20)
